// src/models/types/Location.ts
enum Location {
  AIRSHIP = "AIRSHIP",
  BESAID = "BESAID",
  BESAID_WATERFALL_WAY = "BESAID_WATERFALL_WAY",
  BIKANEL = "BIKANEL",
  CACTUAR_NATION = "CACTUAR_NATION",
  CALM_LANDS = "CALM_LANDS",
  CAVERN_OF_THE_STOLEN_FAYTH = "CAVERN_OF_THE_STOLEN_FAYTH",
  DJOSE = "DJOSE",
  DJOSE_HIGHROAD = "DJOSE_HIGHROAD",
  HOME = "HOME",
  INSIDE_SIN = "INSIDE_SIN",
  KILIKA = "KILIKA",
  KILIKA_WOODS = "KILIKA_WOODS",
  LAKE_MACALANIA = "LAKE_MACALANIA",
  LUCA = "LUCA",
  MACALANIA = "MACALANIA",
  MACALANIA_WOODS = "MACALANIA_WOODS",
  MI_IHEN_HIGHROAD = "MI_IHEN_HIGHROAD",
  MI_IHEN_HIGHROAD_OLDROAD = "MI_IHEN_HIGHROAD_OLDROAD",
  MONT_GAGAZET = "MONT_GAGAZET",
  MONT_GAGAZET_CAVE = "MONT_GAGAZET_CAVE",
  MONT_GAGAZET_MOUNTAINS = "MONT_GAGAZET_MOUNTAINS",
  MOONFLOW = "MOONFLOW",
  MUSHROOM_ROCK_ROAD = "MUSHROOM_ROCK_ROAD",
  OMEGA_RUINS = "OMEGA_RUINS",
  THUNDER_PLAINS = "THUNDER_PLAINS",
  ZANARKAND = "ZANARKAND",
}

type MainLocationKey =
  | Location.BESAID
  | Location.BIKANEL
  | Location.CALM_LANDS
  | Location.CAVERN_OF_THE_STOLEN_FAYTH
  | Location.DJOSE
  | Location.INSIDE_SIN
  | Location.KILIKA
  | Location.LUCA
  | Location.MI_IHEN_HIGHROAD
  | Location.MACALANIA
  | Location.MONT_GAGAZET
  | Location.MUSHROOM_ROCK_ROAD
  | Location.OMEGA_RUINS
  | Location.THUNDER_PLAINS
  | Location.ZANARKAND;

interface MainLocation {
  key: MainLocationKey;
  subLocations?: Location[];
}

export type { MainLocation, MainLocationKey };

export { Location };